requirements.txt
# Core ML dependencies
scikit-learn==1.4.0
xgboost==2.0.3
pandas==2.2.1
numpy==1.26.4

# MLflow and model serving
mlflow==2.10.2
flask==2.3.3
gunicorn==21.2.0

# Data processing and monitoring
python-dotenv==1.0.1
evidently==0.4.0
prometheus-client==0.19.0

# Testing and development
pytest==8.0.2
black==24.2.0
flake8==7.0.0

# API and web
requests==2.31.0
python-dateutil==2.8.2

optuna==3.5.0
imbalanced-learn==0.12.0
plotly==5.18.0
scipy==1.11.4
httpx==0.27.0
python-multipart==0.0.9
kaggle==1.6.0 